requirements_enhanced.txt
# Enhanced requirements with advanced tools

# Existing requirements
agno
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
requests

# PDF and Document Processing
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.21

# Code Analysis and Quality
bandit>=1.7.5
pylint>=2.17.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
safety>=2.3.0

# System Monitoring
psutil>=5.9.0
py-cpuinfo>=9.0.0
GPUtil>=1.4.0

# Advanced Web Scraping
beautifulsoup4>=4.12.0
selenium>=4.11.0
scrapy>=2.10.0
requests-html>=0.10.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Tools
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0

# Development Tools
jupyter>=1.0.0
ipython>=8.14.0
pre-commit>=3.3.0
pytest>=7.4.0
coverage>=7.2.0

# Database Tools
alembic>=1.11.0
redis>=4.6.0
elasticsearch>=8.8.0

# Communication Tools
slack-sdk>=3.21.0
discord.py>=2.3.0
twilio>=8.5.0

# File Processing
watchdog>=3.0.0
send2trash>=1.8.2
pathvalidate>=3.1.0

# Network Tools
httpx>=0.24.0
aiohttp>=3.8.0
websockets>=11.0.0

# Configuration Management
pydantic>=2.0.0
pydantic-settings>=2.0.0
click>=8.1.0
typer>=0.9.0

# Deployment Tools
docker>=6.1.0
kubernetes>=27.2.0
terraform>=1.5.0

# Monitoring and Logging
prometheus-client>=0.17.0
grafana-api>=1.0.3
loguru>=0.7.0
structlog>=23.1.0